EES-3864 - trim down run_tests.py
