+++ tests/robot-tests/run_tests.py
@@ -8,20 +8,24 @@ Run 'python run_tests.py -h' to see argument options
 
 import argparse
 import datetime
-import json
 import os
 import shutil
 from pathlib import Path
 
-import requests
 from dotenv import load_dotenv
 from pabot.pabot import main as pabot_run_cli
 from robot import rebot_cli as robot_rebot_cli
 from robot import run_cli as robot_run_cli
 from scripts.get_webdriver import get_webdriver
+from tests.libs.admin_utilities import (
+    create_test_theme,
+    create_test_topic,
+    delete_test_topic,
+    get_test_themes,
+    setup_authentication,
+)
 from tests.libs.create_emulator_release_files import ReleaseFilesGenerator
 from tests.libs.logger import get_logger
-from tests.libs.setup_auth_variables import setup_auth_variables
 from tests.libs.slack import send_slack_report
 
 current_dir = Path(__file__).absolute().parent
@@ -41,113 +45,6 @@ if pythonpath:
 else:
     os.environ["PYTHONPATH"] = str(current_dir)
 
-
-# Parse arguments
-parser = argparse.ArgumentParser(
-    prog="pipenv run python run_tests.py",
-    description="Use this script to run the UI tests, locally or as part of the CI pipeline, against the environment of your choosing",
-)
-parser.add_argument(
-    "-b",
-    "--browser",
-    dest="browser",
-    default="chrome",
-    choices=["chrome", "firefox", "ie"],
-    help="name of the browser you wish to run the tests with (NOTE: Only chromedriver is automatically installed!)",
-)
-parser.add_argument(
-    "-i",
-    "--interp",
-    dest="interp",
-    default="pabot",
-    choices=["pabot", "robot"],
-    help="interpreter to use to run the tests",
-)
-parser.add_argument(
-    "--processes", dest="processes", help="how many processes should be used when using the pabot interpreter"
-)
-parser.add_argument(
-    "-e",
-    "--env",
-    dest="env",
-    default="test",
-    choices=["local", "dev", "test", "preprod", "prod", "ci"],
-    help="the environment to run the tests against",
-)
-parser.add_argument(
-    "-f",
-    "--file",
-    dest="tests",
-    metavar="{file/dir}",
-    default="tests/",
-    help="test suite or folder of tests suites you wish to run",
-)
-parser.add_argument(
-    "-t", "--tags", dest="tags", nargs="?", metavar="{tag(s)}", help="specify tests you wish to run by tag"
-)
-parser.add_argument(
-    "-v", "--visual", dest="visual", action="store_true", help="display browser window that the tests run in"
-)
-parser.add_argument(
-    "--ci", dest="ci", action="store_true", help="specify that the test are running as part of the CI pipeline"
-)
-parser.add_argument(
-    "--chromedriver",
-    dest="chromedriver_version",
-    metavar="{version}",
-    help="specify which version of chromedriver to use",
-)
-parser.add_argument(
-    "--disable-teardown",
-    dest="disable_teardown",
-    help="disable tearing down of any test data after completion",
-    action="store_true",
-)
-parser.add_argument(
-    "--rerun-failed-tests",
-    dest="rerun_failed_tests",
-    action="store_true",
-    help="rerun individual failed tests and merge results into original run results",
-)
-parser.add_argument(
-    "--rerun-failed-suites",
-    dest="rerun_failed_suites",
-    action="store_true",
-    help="rerun failed test suites and merge results into original run results",
-)
-parser.add_argument(
-    "--print-keywords",
-    dest="print_keywords",
-    action="store_true",
-    help="choose to print out keywords as they are started",
-)
-parser.add_argument("--enable-slack", dest="enable_slack", action="store_true")
-parser.add_argument(
-    "--prompt-to-continue",
-    dest="prompt_to_continue",
-    action="store_true",
-    help="get prompted to continue with test execution upon a failure",
-)
-parser.add_argument("--fail-fast", dest="fail_fast", action="store_true", help="stop test execution on failure")
-parser.add_argument(
-    "--custom-env", dest="custom_env", default=None, help="load a custom .env file (must be in ~/robot-tests directory)"
-)
-
-"""
-NOTE(mark): The slack webhook url, and admin and analyst passwords to access to Admin app are
-stored in the CI pipeline as secret variables, which means they cannot be accessed as normal
-environment variables, and instead must be passed as an argument to this script.
-"""
-parser.add_argument("--slack-webhook-url", dest="slack_webhook_url", default=None, help="URL for Slack webhook")
-parser.add_argument("--admin-pass", dest="admin_pass", default=None, help="manually specify the admin password")
-parser.add_argument("--analyst-pass", dest="analyst_pass", default=None, help="manually specify the analyst password")
-args = parser.parse_args()
-
-if args.custom_env:
-    load_dotenv(args.custom_env)
-else:
-    load_dotenv(".env." + args.env)
-
 assert os.getenv("TIMEOUT") is not None
 assert os.getenv("IMPLICIT_WAIT") is not None
 assert os.getenv("PUBLIC_URL") is not None
@@ -161,6 +58,161 @@ assert os.getenv("WAIT_SMALL") is not None
 assert os.getenv("FAIL_TEST_SUITES_FAST") is not None
 assert os.getenv("IDENTITY_PROVIDER") is not None
 assert os.getenv("WAIT_MEMORY_CACHE_EXPIRY") is not None
+
+# Parse arguments
+parser = argparse.ArgumentParser(
+    prog="pipenv run python run_tests.py",
+    description="Use this script to run the UI tests, locally or as part of the CI pipeline, against the environment of your choosing",
+)
+parser.add_argument(
+    "-b",
+    "--browser",
+    dest="browser",
+    default="chrome",
+    choices=["chrome", "firefox", "ie"],
+    help="name of the browser you wish to run the tests with (NOTE: Only chromedriver is automatically installed!)",
+    required=False,
+)
+parser.add_argument(
+    "-i",
+    "--interp",
+    dest="interp",
+    default="pabot",
+    choices=["pabot", "robot"],
+    help="interpreter to use to run the tests",
+    required=True,
+)
+parser.add_argument(
+    "--processes",
+    dest="processes",
+    help="how many processes should be used when using the pabot interpreter",
+    required=False,
+)
+
+parser.add_argument(
+    "-e",
+    "--env",
+    dest="env",
+    default="test",
+    choices=["local", "dev", "test", "preprod", "prod", "ci"],
+    help="the environment to run the tests against",
+    required=True,
+)
+parser.add_argument(
+    "-f",
+    "--file",
+    dest="tests",
+    metavar="{file/dir}",
+    default="tests/",
+    help="test suite or folder of tests suites you wish to run",
+    required=True,
+)
+parser.add_argument(
+    "-t",
+    "--tags",
+    dest="tags",
+    nargs="?",
+    metavar="{tag(s)}",
+    help="specify tests you wish to run by tag",
+    required=False,
+)
+
+parser.add_argument(
+    "-v",
+    "--visual",
+    dest="visual",
+    action="store_true",
+    help="display browser window that the tests run in",
+    required=False,
+)
+parser.add_argument(
+    "--ci",
+    dest="ci",
+    action="store_true",
+    help="specify that the test are running as part of the CI pipeline",
+    required=False,
+)
+parser.add_argument(
+    "--chromedriver",
+    dest="chromedriver_version",
+    metavar="{version}",
+    help="specify which version of chromedriver to use",
+    required=False,
+)
+parser.add_argument(
+    "--disable-teardown",
+    dest="disable_teardown",
+    help="disable tearing down of any test data after completion",
+    action="store_true",
+    required=False,
+)
+parser.add_argument(
+    "--rerun-failed-tests",
+    dest="rerun_failed_tests",
+    action="store_true",
+    help="rerun individual failed tests and merge results into original run results",
+    required=False,
+)
+parser.add_argument(
+    "--rerun-failed-suites",
+    dest="rerun_failed_suites",
+    action="store_true",
+    help="rerun failed test suites and merge results into original run results",
+    required=False,
+)
+parser.add_argument(
+    "--print-keywords",
+    dest="print_keywords",
+    action="store_true",
+    help="choose to print out keywords as they are started",
+    required=False,
+)
+
+parser.add_argument("--enable-slack", dest="enable_slack", action="store_true", required=False)
+
+parser.add_argument(
+    "--prompt-to-continue",
+    dest="prompt_to_continue",
+    action="store_true",
+    help="get prompted to continue with test execution upon a failure",
+    required=False,
+)
+
+parser.add_argument(
+    "--fail-fast", dest="fail_fast", action="store_true", help="stop test execution on failure", required=False
+)
+
+parser.add_argument(
+    "--custom-env",
+    dest="custom_env",
+    default=None,
+    help="load a custom .env file (must be in ~/robot-tests directory)",
+    required=False,
+)
+
+"""
+NOTE(mark): The slack webhook url, and admin and analyst passwords to access to Admin app are
+stored in the CI pipeline as secret variables, which means they cannot be accessed as normal
+environment variables, and instead must be passed as an argument to this script.
+"""
+parser.add_argument(
+    "--slack-webhook-url", dest="slack_webhook_url", default=None, help="URL for Slack webhook", required=False
+)
+
+parser.add_argument(
+    "--admin-pass", dest="admin_pass", default=None, help="manually specify the admin password", required=False
+)
+
+parser.add_argument(
+    "--analyst-pass", dest="analyst_pass", default=None, help="manually specify the analyst password", required=False
+)
+
+args = parser.parse_args()
+
+if args.custom_env:
+    load_dotenv(args.custom_env)
+else:
+    load_dotenv(".env." + args.env)
 
 
 if args.slack_webhook_url:
@@ -218,101 +270,9 @@ if args.ci:
     robotArgs += ["--removekeywords", "name:common.user goes to url"]  # To hide basic auth credentials
 
 
-def admin_request(method, endpoint, body=None):
-    assert method and endpoint
-    assert os.getenv("ADMIN_URL") is not None
-    assert os.getenv("IDENTITY_LOCAL_STORAGE_ADMIN") is not None
-
-    if method == "POST":
-        assert body is not None, "POST requests require a body"
-
-    requests.sessions.HTTPAdapter(pool_connections=50, pool_maxsize=50, max_retries=3)
-    session = requests.Session()
-
-    # To prevent InsecureRequestWarning
-    requests.packages.urllib3.disable_warnings()
-
-    jwt_token = json.loads(os.getenv("IDENTITY_LOCAL_STORAGE_ADMIN"))["access_token"]
-    headers = {
-        "Content-Type": "application/json",
-        "Authorization": f"Bearer {jwt_token}",
-    }
-    response = session.request(
-        method, url=f'{os.getenv("ADMIN_URL")}{endpoint}', headers=headers, stream=True, json=body, verify=False
-    )
-
-    if response.status_code in {401, 403}:
-        logger.info("Attempting re-authentication...")
-
-        # Delete identify files and re-attempt to fetch them
-        setup_authentication(clear_existing=True)
-        jwt_token = json.loads(os.environ["IDENTITY_LOCAL_STORAGE_ADMIN"])["access_token"]
-        response = session.request(
-            method,
-            url=f'{os.getenv("ADMIN_URL")}{endpoint}',
-            headers={
-                "Content-Type": "application/json",
-                "Authorization": f"Bearer {jwt_token}",
-            },
-            stream=True,
-            json=body,
-            verify=False,
-        )
-
-        assert response.status_code not in {401, 403}, "Failed to reauthenticate."
-
-    assert response.status_code < 300, f"Admin request responded with {response.status_code} and {response.text}"
-    return response
-
-
-def get_test_themes():
-    return admin_request("GET", "/api/themes")
-
-
-def create_test_theme():
-    return admin_request("POST", "/api/themes", {"title": "Test theme", "summary": "Test theme summary"})
-
-
-def create_test_topic():
-    assert os.getenv("TEST_THEME_ID") is not None
-
-    topic_name = f'UI test topic {os.getenv("RUN_IDENTIFIER")}'
-    resp = admin_request("POST", "/api/topics", {"title": topic_name, "themeId": os.getenv("TEST_THEME_ID")})
-
-    os.environ["TEST_TOPIC_NAME"] = topic_name
-    os.environ["TEST_TOPIC_ID"] = resp.json()["id"]
-
-
-def delete_test_topic():
-    if os.getenv("TEST_TOPIC_ID") is not None:
-        admin_request("DELETE", f'/api/topics/{os.getenv("TEST_TOPIC_ID")}')
-
-
-def setup_authentication(clear_existing=False):
-    # Don't need BAU user if running general_public tests
-    if "general_public" not in args.tests:
-        setup_auth_variables(
-            user="ADMIN",
-            email=os.getenv("ADMIN_EMAIL"),
-            password=os.getenv("ADMIN_PASSWORD"),
-            clear_existing=clear_existing,
-            identity_provider=os.getenv("IDENTITY_PROVIDER"),
-        )
-
-    # Don't need analyst user if running admin/bau or admin_and_public/bau tests
-    if f"{os.sep}bau" not in args.tests:
-        setup_auth_variables(
-            user="ANALYST",
-            email=os.getenv("ANALYST_EMAIL"),
-            password=os.getenv("ANALYST_PASSWORD"),
-            clear_existing=clear_existing,
-            identity_provider=os.getenv("IDENTITY_PROVIDER"),
-        )
-
-
 # Auth not required with general_public tests
 if args.tests and "general_public" not in args.tests:
-    setup_authentication()
+    setup_authentication(args)
 
     # NOTE(mark): Tests that alter data only occur on local and dev environments
     if args.env in ["local", "dev"]:
